Revert "Added the home page layout"
