+++ src/css/style.css
@@ -40,11 +40,6 @@ body {
   background-color: aqua;
 }
 
-.home-col {
-  height: 300px; 
-  margin-right: 50px;
-  margin-left: 20px;
-}
 /* footer */
 .footer {
   text-align: center;
