+++ src/index.html
@@ -28,9 +28,6 @@
         </div>
       </div>
     </nav>
-
-
-    <div id="home-page"><div id="row" class="row"></div></div>
     <div id="display-movies"></div>
   </body>
 </html>
